some cleaning up
