+++ src/Simple.Web.Razor.Tests/MockHandler.cs
@@ -2,7 +2,6 @@ namespace Simple.Web.Razor.Tests
 {
     using System;
     using System.Collections.Generic;
-    using System.Dynamic;
     using System.Linq;
     using Links;
     using MediaTypeHandling;
