+++ src/Simple.Web.Razor.Tests/RazorHtmlContentTypeHandlerTests.cs
@@ -1,6 +1,5 @@
 namespace Simple.Web.Razor.Tests
 {
-    using System;
     using System.IO;
     using Xunit;
