+++ src/Simple.Web.Razor.Tests/RazorTypeBuilderTests.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Simple.Web.Razor.Tests
 {
-    using System.IO;
     using Xunit;
 
     public class RazorTypeBuilderTests
